Add healthcheck to DB container
Without this, sometimes when building the containers, the application container
would struggle to connect to the database and therefore failed with "Connection
Refused" when trying to update the DB schema or seed fixtures.

This health check should alleviate this issue, by making the application depend
on the DB, which now has a healthcheck set on it.