Fix issue with test translation dependencies
Essentially, when you run the seed command for the test environment, it was
seeding test data and non-test data because of the tests classes
inheritance setup. Because they were inheriting their parent classes, the parent
classes had their dependencies set to non-test fixtures. So when we run the seed
command for test fixtures, it was including non-test fixtures too.