Amend Repositories to not extend ServiceEntityRepository
Rather than extending ServiceEntityRepository and being heavily coupled to
Doctrine, I've injected the EntityManagerInterface instead and introduced
more explicit methods (rather than relying on out of the box methods that
Doctrine supplies).

This isn't perfect, because we're still heavily coupled to Doctrine, but it is
a step in the right direction into decoupling the repositories from Doctrine.