Amend `UnitTestCase` so that MockeryPHPUnitIntegration trait used
This `Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration` trait does lots of
great things. The main thing I want to use it for, is after every test it will
close down Mockery, so that I do not have to call it in a `tearDown` method in
every test.

Interestingly, this actually highlighted some issues with Mockery assertions in
my current unit tests, so it was well worth adding this in.